refactor(setup): split ReadinessSteps into one file per step
Each exported step now lives in its own steps/ file behind a barrel;
StepProps and CONTROL_STYLE move to steps/types. config.tsx imports the
barrel instead of ReadinessSteps. Pure code motion.
