+++ src/modules/SetupWalkthrough/steps/index.ts
@@ -1,0 +1,15 @@
+/**
+ * Readiness steps barrel
+ *
+ * One file per walkthrough step; `config.tsx` wires them into STEP_CONFIGS.
+ */
+export { BasicsStep } from "./BasicsStep";
+export { GoalStep } from "./GoalStep";
+export { OrganizationStep } from "./OrganizationStep";
+export { ReadyStep } from "./ReadyStep";
+export { SetupOperationError } from "./SetupOperationError";
+export { SharingStep } from "./SharingStep";
+export { ToolsStep } from "./ToolsStep";
+export { TutorialStep } from "./TutorialStep";
+export { WorkModelStep } from "./WorkModelStep";
+export { CONTROL_STYLE, type StepProps } from "./types";
